+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.0.7
+
+Fixed [`positioned` subproperties are considered duplicates for `execute-duplicate`](https://github.com/TheSalts/datapack-optimization-helper/issues/4)
+
+## 1.0.6
+
+Fixed [Wrong warning (function group)](https://github.com/TheSalts/datapack-optimization-helper/issues/3)
+
 ## 1.0.5
 
 Fixed [Incorrect warning with execute in](https://github.com/TheSalts/datapack-optimization-helper/issues/2)